requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
bcrypt==4.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
Mako==1.3.7
MarkupSafe==3.0.2
passlib==1.7.4
pip==24.3.1
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
pydantic-settings==2.6.1
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==65.5.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
gunicorn==20.1.0
pywin32==305; platform_system == "Windows"  
